Add files via upload
Visualization helped to understand that Age is an important variable, also it correlated with years of experience, which makes sense as with age a person gains experience. We also noticed that below age 30, males cause more losses than females, singles perform poorly than married & single males are the worst performers.